requirements.txt
Django>=5.0,<5.2
django-environ>=0.11.2
requests>=2.31.0
urllib3>=2.0.0
yfinance>=0.2.40
finnhub-python>=2.4.20
alpha-vantage>=2.3.1
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.4.0
tensorflow>=2.15.0
ta>=0.11.0
textblob>=0.18.0
plotly>=5.18.0
redis>=5.0.0
django-redis>=5.4.0
tenacity>=8.2.3
ratelimit>=2.2.1
python-dateutil>=2.8.2
gunicorn>=21.2.0
whitenoise>=6.6.0
